+++ packages/context/src/metadata/metadata.keys.ts
@@ -7,3 +7,6 @@ export const BEAN_NAME_METADATA_KEY = Symbol("Bean-Name");
 // CONTROLLER
 export const CONTROLLER_PATH_METADATA_KEY = Symbol("Controller-Path");
 export const CONTROLLER_VERSION_METADATA_KEY = Symbol("Controller-Version");
+
+// BEAN
+export const LIFECYCLE_TYPE_METADATA_KEY = "__Lifecycle-Type";