+++ servers/express-server/src/server/ExpressServer.ts
@@ -75,7 +75,8 @@ export class ExpressServer extends BaseServer<express.Application, express.Appli
                     this.logger.info(`======= ENV: ${context.applicationOptions.environment} =======`);
                     this.logger.info(`🚀 App listening on the port ${context.applicationOptions.port}`);
                     this.logger.info(`=================================`);
-
+                    // mark the server as started
+                    super.started();
                     // Server initialized
                     resolve(this.appView());
                 });