+++ packages/authorization/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nodeboot/authorization
 
+## 1.1.10
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.6.0
+    -   @nodeboot/engine@1.1.10
+
 ## 1.1.9
 
 ### Patch Changes