+++ packages/core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nodeboot/core
 
+## 1.8.0
+
+### Minor Changes
+
+-   Implement lifecycle management for Application Feature Adapters
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.6.0
+    -   @nodeboot/config@1.1.8
+    -   @nodeboot/di@1.1.8
+    -   @nodeboot/engine@1.1.10
+
 ## 1.7.0
 
 ### Minor Changes