+++ starters/firebase/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @nodeboot/starter-firebase
 
+## 1.0.14
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.6.0
+    -   @nodeboot/core@1.8.0
+    -   @nodeboot/config@1.1.8
+    -   @nodeboot/engine@1.1.10
+
 ## 1.0.13
 
 ### Patch Changes