+++ starters/persistence/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nodeboot/starter-persistence
 
+## 1.7.0
+
+### Minor Changes
+
+-   Implement lifecycle management for Application Feature Adapters
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.6.0
+    -   @nodeboot/core@1.8.0
+    -   @nodeboot/config@1.1.8
+    -   @nodeboot/di@1.1.8
+
 ## 1.6.6
 
 ### Patch Changes