+++ tests/jest-koa-demo/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @nodeboot/jest-koa-demo
 
+## 1.0.19
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/starter-persistence@1.7.0
+    -   @nodeboot/koa-server@1.4.0
+    -   @nodeboot/context@1.6.0
+    -   @nodeboot/core@1.8.0
+    -   @nodeboot/authorization@1.1.10
+    -   @nodeboot/config@1.1.8
+    -   @nodeboot/di@1.1.8
+    -   @nodeboot/starter-actuator@1.1.12
+    -   @nodeboot/starter-openapi@1.7.6
+
 ## 1.0.18
 
 ### Patch Changes